requirements.txt
robotframework
python-can
pyserial
ruff